backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
attrs==25.3.0
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
datasets==4.0.0
dill==0.3.8
filelock==3.18.0
Flask==3.1.1
flask-cors==6.0.0
Flask-SQLAlchemy==3.1.1
frozenlist==1.7.0
fsspec==2025.3.0
greenlet==3.2.4
hf-xet==1.1.7
html2text==2025.4.15
huggingface-hub==0.34.4
idna==3.10
itsdangerous==2.2.0
jieba==0.42.1
Jinja2==3.1.6
loguru==0.7.3
lxml==6.0.0
MarkupSafe==3.0.2
multidict==6.6.4
multiprocess==0.70.16
numpy==2.3.2
packaging==25.0
pandas==2.3.1
propcache==0.3.2
pyahocorasick==2.2.0
pyarrow==21.0.0
pycorrector==1.1.3
pypinyin==0.55.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
pytz==2025.2
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.4
safetensors==0.6.2
six==1.17.0
SQLAlchemy==2.0.41
tokenizers==0.21.4
tqdm==4.67.1
transformers==4.55.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
xxhash==3.5.0
yarl==1.20.1
